refactored clang files to improve readability/sustainability
