+++ bin/clang/computeLoops.h
@@ -42,27 +42,35 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 Questions? Contact sst-macro-help@sandia.gov
 */
 
-#ifndef bin_clang_astconsumers_h
-#define bin_clang_astconsumers_h
+#ifndef bin_clang_compute_loops_h
+#define bin_clang_compute_loops_h
 
-#include "clangHeaders.h"
-#include "astVisitor.h"
-#include "globalVarNamespace.h"
+struct Loop {
+  struct Body {
+    int depth;
+    int flops;
+    int intops;
+    int writeBytes;
+    int readBytes;
+    std::list<Loop> subLoops;
+    Body() : flops(0), intops(0), readBytes(0), writeBytes(0) {}
+  };
 
-class ReplaceASTConsumer : public clang::ASTConsumer {
- public:
-  ReplaceASTConsumer(clang::Rewriter &R, ReplGlobalASTVisitor& r) :
-    visitor_(r)
-  {
+  std::string tripCount;
+  Body body;
+  Loop(int depth){
+    body.depth = depth;
   }
-
-  bool HandleTopLevelDecl(clang::DeclGroupRef DR) override;
-
- private:
-  ReplGlobalASTVisitor& visitor_;
-
 };
 
-
+struct ForLoopSpec
+{
+  bool increment = true;
+  clang::Expr* stride;
+  clang::Expr* init;
+  clang::Expr* predicateMax;
+  clang::NamedDecl* incrementer;
+};
 
 #endif
+
