+++ bin/clang/dataFlow.h
@@ -41,28 +41,94 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 Questions? Contact sst-macro-help@sandia.gov
 */
+#ifndef bin_clang_dataFlow_h
+#define bin_clang_dataFlow_h
 
-#ifndef bin_clang_astconsumers_h
-#define bin_clang_astconsumers_h
+struct DataFlowMap {
+  struct Access {
+    uint32_t generation;
+  };
 
-#include "clangHeaders.h"
-#include "astVisitor.h"
-#include "globalVarNamespace.h"
+  std::map<clang::NamedDecl*,uint32_t> generation;
 
-class ReplaceASTConsumer : public clang::ASTConsumer {
- public:
-  ReplaceASTConsumer(clang::Rewriter &R, ReplGlobalASTVisitor& r) :
-    visitor_(r)
-  {
+  std::set<uint64_t> accessed;
+  static void
+  append(uint64_t& id, uint16_t modifier, uint16_t idx){
+    id |= (modifier << idx);
   }
-
-  bool HandleTopLevelDecl(clang::DeclGroupRef DR) override;
-
- private:
-  ReplGlobalASTVisitor& visitor_;
-
+  uint16_t id_count;
+  DataFlowMap() : id_count(0) {}
 };
 
 
+struct MemoryLocation {
+  char label[128];
+
+  template <class T>
+  void
+  append(const T& t){
+    //::memcpy(&label[size], &t, sizeof(T));
+    T* arr = (T*) &label[size];
+    *arr = t;
+    size += sizeof(T);
+  }
+
+  const char* c_str() const {
+    char* ret = new char[128];
+    int writeIdx = 0;
+    for (int i=0; i < size; ++i){
+      if (label[i] != '\0'){
+        ret[writeIdx++] = label[i];
+      }
+    }
+    ret[writeIdx] = '\0';
+    return ret;
+  }
+
+  void
+  updateGeneration(uint32_t gen){
+    maxGen = std::max(maxGen, gen);
+  }
+
+  void
+  append(const MemoryLocation& mloc){
+    ::memcpy(&label[size], mloc.label, mloc.size);
+    size += mloc.size;
+    maxGen = std::max(maxGen, mloc.maxGen);
+  }
+
+  uint32_t maxGen;
+  uint8_t size;
+  MemoryLocation() : maxGen(0), size(0) {}
+};
+
+struct MemoryLocationCompare {
+  bool operator()(const MemoryLocation& lhs, const MemoryLocation& rhs) const {
+    if (lhs.size != rhs.size) return lhs.size < rhs.size;
+    int cmp = memcmp(lhs.label, rhs.label, lhs.size);
+    return cmp < 0;
+  }
+};
+
+struct AccessHistory {
+  uint32_t lastReadGeneration;
+  uint32_t lastWriteGeneration;
+
+  AccessHistory() : lastReadGeneration(0), lastWriteGeneration(0)
+  {
+  }
+
+  bool newAccess(uint32_t maxGenDependence, uint32_t currentGeneration, bool isLHS){
+    if (isLHS){
+      bool updated = lastWriteGeneration < maxGenDependence || lastWriteGeneration == 0;
+      lastWriteGeneration = currentGeneration;
+      return updated;
+    } else {
+      bool updated = lastReadGeneration < maxGenDependence || lastReadGeneration == 0;
+      lastReadGeneration = currentGeneration;
+      return updated;
+    }
+  }
+};
 
 #endif
